requirements.txt
python-vlc==3.0.12118
